+++ src/emulator/hardware.rs
@@ -217,7 +217,7 @@ impl Hardware {
 
     fn read_timer(&self) -> u16 {
         let memory = self.memory.borrow();
-        (memory.read_halfword(THR) as u16) << 8 | memory.read_halfword(TLR) as u16
+        memory.read_halfword(THR) << 8 | memory.read_halfword(TLR)
     }
 
     fn write_timer(&self, value: u16) {
